LayoutTests/fast/events/mac/mouse-down-can-start-selection-quirk-expected.txt

PASS getSelection().toString() is ""
PASS sliderPosition is 180
PASS isDraggingSlider is false
PASS successfullyParsed is true

TEST COMPLETE
 Hello world
